+++ src/actions/frequencies.js
@@ -105,9 +105,53 @@ export const updateFrequencyData = (dispatch, getState) => {
 };
 
 /* debounce works better than throttle, as it _won't_ update while events are still coming in (e.g. dragging the date slider) */
-export const updateFrequencyDataDebounced = debounce(updateFrequencyData, 500, { leading: false, trailing: true });
+export const updateFrequencyDataDebounced = debounce(
+  updateFrequencyData, 500, { leading: false, trailing: true }
+);
 
 export const toggleNormalization = (dispatch, getState) => {
   dispatch({type: types.TOGGLE_FREQUENCY_NORMALIZATION});
   updateFrequencyData(dispatch, getState);
+};
+
+export const processFrequenciesJSON = (rawJSON, tree, controls) => {
+  /* this function can throw */
+  const normaliseData = false;
+  const pivots = rawJSON.pivots.map((d) => Math.round(parseFloat(d) * 100) / 100);
+  const ticks = [pivots[0]];
+  const tick_step = (pivots[pivots.length - 1] - pivots[0]) / 6 * 10 / 10;
+  while (ticks[ticks.length - 1] < pivots[pivots.length - 1]) {
+    ticks.push((ticks[ticks.length - 1] + tick_step) * 10 / 10);
+  }
+  if (!tree.loaded) {
+    throw new Error("tree not loaded");
+  }
+  const data = [];
+  tree.nodes.filter((d) => !d.hasChildren).forEach((n) => {
+    if (!rawJSON[n.strain]) {
+      console.warn(`No tip frequency information for ${n.strain}`);
+      return;
+    }
+    data.push({
+      idx: n.arrayIdx,
+      values: rawJSON[n.strain].frequencies,
+      weight: rawJSON[n.strain].weight
+    });
+  });
+  const matrix = computeMatrixFromRawData(
+    data,
+    normaliseData,
+    pivots,
+    tree.nodes,
+    tree.visibility,
+    controls.colorScale,
+    controls.colorBy
+  );
+  return {
+    data,
+    pivots,
+    ticks,
+    matrix,
+    normaliseData
+  };
 };
