+++ src/actions/types.js
@@ -1,6 +1,5 @@
 export const UPDATE_TIP_RADII = "UPDATE_TIP_RADII";
 export const NEW_COLORS = "NEW_COLORS";
-export const RECEIVE_TREE = "RECEIVE_TREE";
 export const INITIALISE_FREQUENCIES = "INITIALISE_FREQUENCIES";
 export const FREQUENCY_MATRIX = "FREQUENCY_MATRIX";
 export const TOGGLE_FREQUENCY_NORMALIZATION = "TOGGLE_FREQUENCY_NORMALIZATION";
@@ -19,12 +18,10 @@ export const CHANGE_DATES_VISIBILITY_THICKNESS = "CHANGE_DATES_VISIBILITY_THICKN
 export const CHANGE_ABSOLUTE_DATE_MIN = "CHANGE_ABSOLUTE_DATE_MIN";
 export const CHANGE_ABSOLUTE_DATE_MAX = "CHANGE_ABSOLUTE_DATE_MAX";
 export const CHANGE_GEO_RESOLUTION = "CHANGE_GEO_RESOLUTION";
-export const NEW_DATASET = "NEW_DATASET";
 export const CLEAN_START = "CLEAN_START";
 export const APPLY_FILTER = "APPLY_FILTER";
 export const UPDATE_VISIBILITY_AND_BRANCH_THICKNESS = "UPDATE_VISIBILITY_AND_BRANCH_THICKNESS";
 export const TOGGLE_MUT_TYPE = "TOGGLE_MUT_TYPE";
-export const ANALYSIS_SLIDER = "ANALYSIS_SLIDER";
 export const CHANGE_ANALYSIS_VALUE = "CHANGE_ANALYSIS_VALUE";
 export const CHANGE_ANIMATION_START = "CHANGE_ANIMATION_START";
 export const CHANGE_ANIMATION_TIME = "CHANGE_ANIMATION_TIME";
